Don't install cythonize dependencies unless cythonizing
Numpy is not building in my environment, but I should not need to have
it if all I want to build is the default libvmaf target.

Additionally:

Instead of using the mtime of the .venv directory as a marker for
whether meson and ninja were successfully installed, use their wrapper
scripts from .venv/bin.

The $(VENV) target uses set -e and escaped newlines to run all commands
in a single shell invocation. Unless these commands are modifying the
shell's environment, this is redundant and mimics the way make already
works when you put a single command on each line.